requirements.txt
steam
dota2
python-dotenv